fixed up redux-persist
Moved store creation out of app.js and into index.js
Moved persist config and persistor out of featuresReducer.js and into configureStore.js
