+++ frontend/src/App.js
@@ -2,11 +2,8 @@
 import React from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 import { Container } from 'react-bootstrap';
-import { Provider } from 'react-redux';
-import { persistStore } from 'redux-persist';
-import { PersistGate } from 'redux-persist/integration/react';
+
 //app imports
-import { myStore } from './store/configureStore';
 import Header from './components/layout/Header';
 import Footer from './components/layout/Footer';
 import HomeScreen from './screens/HomeScreen';
@@ -17,35 +14,28 @@ import ItemListScreen from './screens/ItemListScreen';
 import UsersScreen from './screens/UsersScreen';
 import ItemScreen from './screens/ItemScreen';
 
-const store = myStore();
-const persistor = persistStore(store);
-
 const App = () => {
   return (
-    <Provider store={store}>
-      <PersistGate loading={null} persistor={persistor}>
-        <Router>
-          <Header />
-          <main className='py-3'>
-            <Container>
-              <Route path='/profile' component={ProfileScreen} />
-              <Route path='/login' component={LoginScreen} />
-              <Route path='/register' component={RegisterScreen} />
-              <Route path='/items/:id' component={ItemScreen} exact />
-              <Route path='/items' component={ItemListScreen} exact />
-              <Route
-                path='/items/page/:pageNumber'
-                component={ItemListScreen}
-                exact
-              />
-              <Route path='/users' component={UsersScreen} />
-              <Route path='/' component={HomeScreen} exact />
-            </Container>
-          </main>
-          <Footer />
-        </Router>
-      </PersistGate>
-    </Provider>
+    <Router>
+      <Header />
+      <main className='py-3'>
+        <Container>
+          <Route path='/profile' component={ProfileScreen} />
+          <Route path='/login' component={LoginScreen} />
+          <Route path='/register' component={RegisterScreen} />
+          <Route path='/items/:id' component={ItemScreen} exact />
+          <Route path='/items' component={ItemListScreen} exact />
+          <Route
+            path='/items/page/:pageNumber'
+            component={ItemListScreen}
+            exact
+          />
+          <Route path='/users' component={UsersScreen} />
+          <Route path='/' component={HomeScreen} exact />
+        </Container>
+      </main>
+      <Footer />
+    </Router>
   );
 };
