+++ frontend/src/index.js
@@ -1,12 +1,26 @@
+//package imports
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { Provider } from 'react-redux';
+import { persistStore } from 'redux-persist';
+import { PersistGate } from 'redux-persist/integration/react';
+
+//app imports
 import './bootstrap.min.css';
 import './index.css';
 import App from './App';
+import myStore from './store/configureStore';
+
+const store = myStore();
+let persistor = persistStore(store);
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <Provider store={store}>
+      <PersistGate loading={null} persistor={persistor}>
+        <App />
+      </PersistGate>
+    </Provider>
   </React.StrictMode>,
   document.getElementById('root')
 );
